Move storage tab to separate file
