+++ components/settings/storage.js
@@ -1,0 +1,78 @@
+var html = require('choo/html')
+var css = require('sheetify')
+
+var backendStyle = css`
+  :host {
+    padding: 10px;
+    padding-right: 15px;
+    border-bottom: 1px solid #999;
+  }
+`
+
+function StorageTab () {
+  return {
+    name: 'storage',
+    description: 'Define data urls and zoom levels for map storage',
+    use: function (emitter) {
+      var self = this
+      emitter.on('settings:storage:url:update', function (index, url) {
+        self.data.backends[index].url = url
+        emitter.emit('settings:dirty')
+      })
+      emitter.on('settings:storage:minzoom:update', function (index, min) {
+        var backend = self.data.backends[index]
+        backend.zoom.min = Math.min(Number(min), backend.zoom.max)
+        emitter.emit('settings:dirty')
+      })
+      emitter.on('settings:storage:maxzoom:update', function (index, max) {
+        var backend = self.data.backends[index]
+        backend.zoom.max = Math.max(Number(max), backend.zoom.min)
+        emitter.emit('settings:dirty')
+      })
+      emitter.on('settings:storage:active:update', function (index) {
+        var backend = self.data.backends[index]
+        backend.active = !backend.active
+        emitter.emit('settings:dirty')
+      })
+    },
+    render: function (emit) {
+      var backends = this.data.backends
+      var content = backends.map(function (item, index) {
+        var zoom = item.zoom
+        return html`<div class=${backendStyle}>
+          <label for='url'>data url</label>
+          <input type='url' name='url' value=${item.url || ''} placeholder='https://example.com' required style='margin-top: 3px; margin-bottom: 10px; width: 100%;' onchange=${(e) => emit('settings:storage:url:update', index, e.target.value)}>
+          <label for='minzoom'>min zoom level (${zoom.min})</label>
+          <input type='range' name='minzoom' min='1' max='21' step='1' value=${zoom.min} style='width: 100%;' onchange=${(e) => emit('settings:storage:minzoom:update', index, e.target.value)}>
+          <label for='maxzoom'>max zoom level (${zoom.max})</label>
+          <input type='range' name='maxzoom' min='1' max='21' step='1' value=${zoom.max} style='width: 100%;' onchange=${(e) => emit('settings:storage:maxzoom:update', index, e.target.value)}>
+          <label for='active'>active</label>
+          <input type='checkbox' name='active' style='margin-left: 10px;' onchange=${(e) => emit('settings:storage:active:update', index)} ${item.active ? 'checked' : ''} value=${item.active ? true : false}>
+        </div>`
+      })
+      return html`<div>${content}</div>`
+    },
+    dirty: false,
+    data: {
+      backends: [
+        {
+          url: 'https://ipfs.io/ipfs/QmVCYUK51Miz4jEjJxCq3bA6dfq5FXD6s2EYp6LjHQhGmh',
+          zoom: { min: 1, max: 21 },
+          active: true
+        },
+        {
+          url: 'https://peermaps.linkping.io',
+          zoom: { min: 5, max: 21 },
+          active: false
+        },
+        {
+          url: 'http://localhost:8000',
+          zoom: { min: 11, max: 18 },
+          active: true
+        }
+      ]
+    }
+  }
+}
+
+module.exports = StorageTab
